Refactor to use new extension method.
