+++ src/OmniSharp.Roslyn.CSharp/Services/Navigation/FindImplementationsService.cs
@@ -34,7 +34,7 @@ namespace OmniSharp.Roslyn.CSharp.Services.Navigation
             {
                 var semanticModel = await document.GetSemanticModelAsync();
                 var sourceText = await document.GetTextAsync();
-                var position = sourceText.Lines.GetPosition(new LinePosition(request.Line, request.Column));
+                var position = sourceText.GetTextPosition(request);
 
                 var quickFixes = new List<QuickFix>();
                 var symbol = await SymbolFinder.FindSymbolAtPositionAsync(semanticModel, position, _workspace);
