+++ src/OmniSharp.Roslyn.CSharp/Services/TestCommands/TestCommandService.cs
@@ -7,6 +7,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Text;
+using OmniSharp.Extensions;
 using OmniSharp.Mef;
 using OmniSharp.Models;
 using OmniSharp.Models.TestCommand;
@@ -38,7 +39,7 @@ namespace OmniSharp.Roslyn.CSharp.Services.TestCommands
                 var semanticModel = await document.GetSemanticModelAsync();
                 var syntaxTree = semanticModel.SyntaxTree;
                 var sourceText = await document.GetTextAsync();
-                var position = sourceText.Lines.GetPosition(new LinePosition(request.Line, request.Column));
+                var position = sourceText.GetTextPosition(request);
                 var node = syntaxTree.GetRoot().FindToken(position).Parent;
 
                 SyntaxNode method = node.FirstAncestorOrSelf<MethodDeclarationSyntax>();
